finish editing comments and cleanup codebase
