--- index.html
@@ -1,12 +1,0 @@
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Redis Rate</title>
-</head>
-
-<body>
-</body>
-
-</html>
