+++ server/LoginModel.js
@@ -1,19 +1,10 @@
 const { Pool } = require('pg');
+require('dotenv').config()
 
-// set up URI, set to process env
-const PG_URI = "postgres://ypokusyr:OJ4Ka4LaJgPRyws7UnhXPG3OgYdNLrhN@ziggy.db.elephantsql.com:5432/ypokusyr";
+const PG_URI = process.env.PG_URI;
 
-// add URI to env file
-
-// create a pool
 const pool = new Pool({
   connectionString: PG_URI
 });
 
-// export our model
 module.exports = pool;
-// module.exports = {
-//   query: (text, params, callback) => {
-//     console.log('executed query', text);
-//     return pool.query(text, params, callback);
-// };
