+++ src/app.jsx
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import ReactDom from 'react-dom';
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
 import Login from './Login.jsx';
@@ -6,18 +6,10 @@ import SignUp from './SignUp.jsx';
 import Connect from './Connect.jsx';
 import Navigation from './Navigation.jsx';
 import ContextProvider from './context/ContextProvider';
-// import Home from './Home.jsx'
-import Memory from './Metrics/Memory.jsx'
 import { createMuiTheme, ThemeProvider, makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { colors } from '@material-ui/core';
-// import palette from './palette';
-// const useStyles = makeStyles(theme => ({
-//   root : { backgroundColor : "pink",
-// }
-// }));
 
-// const classes = useStyles();
 
 const theme = createMuiTheme ({
   palette: {
@@ -28,8 +20,6 @@ const theme = createMuiTheme ({
     
   }
 });
-
-// const theBack = { background: 'linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)' };
 
 
 const mainElement = document.createElement('div');
@@ -53,7 +43,7 @@ const App = () => {
           <Navigation />
         </Route>
       </Switch>
-    </div >
+    </div>
   );
 };
