--- src/app/context/index.ts
@@ -1,5 +1,0 @@
-import React from 'react';
-const AppContext = React.createContext<any>({});
-export default AppContext;
-
-// this page is for storing all the context like a container.
