--- src/app/styles/styles.css
@@ -1,65 +1,0 @@
-
-
-#loginPage {
-  width: 100%;
-  height: 100%;
-  padding: 0;
-  display: block;
-  margin: auto;
-  max-height: 300px;
-  max-width: 300px;
-  align-items: center;
-}
-
-img {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-left: 50px;
-}
-
-#button {
-  margin-top: 12px;
-  padding: 15px;
-  border-radius: 15px;
-}
-
-#butt:hover{
-  /* background-color: #f3b4ca;  */
-  /* opacity: 10%; */
-  color: rgb(139, 226, 226);
-  border: 1px solid rgb(139, 226, 226)
-
-}
-
-#signup {
-  display:flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-#signupForm {
-  display:flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-#form {
-  margin: 15px;
-  display: flex;
-  flex-direction: column;
-}
-
-#MemoryPageTitle {
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-#memoryGraphs {
-  display: flex;
-  flex-direction: column;
-}
-
-#memoryBar {
-  margin-top: 20px;
-}
-
