+++ src/styles/styles.css
@@ -14,15 +14,12 @@
   display: flex;
   flex-direction: column;
   color: rgb(214, 209, 209);
-
 }
 
 #connectPage {
   width: 100%;
   height: 100%;
   padding: 0;
-  /* background: linear-gradient(0deg, rgba(189,217,217,1) 0%, rgba(65,64,62,1) 69%); */
-  display: block;
   margin: auto;
   max-height: 300px;
   max-width: 300px;
@@ -52,8 +49,6 @@
   border-radius: 15px;
 }
 #loginbutt:hover {
-  /* background-color: #f3b4ca;  */
-  /* opacity: 10%; */
   color: rgb(139, 226, 226);
   border: 1px solid rgb(139, 226, 226);
 }
@@ -64,10 +59,6 @@ img {
   align-items: center;
   margin-left: 50px;
 }
-
-/* #texts {
-  color: rgb(214, 209, 209);
-} */
 
 #signup {
   display: flex;
@@ -87,24 +78,6 @@ img {
   justify-content: center;
   align-items: center;
 }
-
-
-
-/* #MemoryPageTitle {
-  font-family: Arial, Helvetica, sans-serif;
-} */
-/* 
-#signupForm {
-
-#memoryGraphs {
-  display: flex;
-  flex-direction: column;
-}
-
-#memoryBar {
-  margin-top: 20px;
-}
-*/
 
 #homeFont {
   font-family: 'Ubuntu', sans-serif;
